+++ server/src/main/java/com/projectdarkstar/rpc/server/ServerCallbackCache.java
@@ -14,29 +14,29 @@ import java.io.Serializable;
 public class ServerCallbackCache implements CallbackCache, Serializable {
     private static final long serialVersionUID = 1L;
 
-    private static class RemoteCallMap extends ScalableHashMap<Long, ManagedSerializable<RemoteCall>> {
+    private static class RemoteCallMap extends ScalableHashMap<Integer, ManagedSerializable<RemoteCall>> {
         private static final long serialVersionUID = 1L;
     }
 
-    private final ManagedReference<ManagedSerializable<Long>> nextRequestId;
+    private final ManagedReference<ManagedSerializable<Integer>> nextRequestId;
     private final ManagedReference<RemoteCallMap> callbacks;
 
     ServerCallbackCache() {
         final DataManager dataManager = AppContext.getDataManager();
-        this.nextRequestId = dataManager.createReference(new ManagedSerializable<Long>(1L));
+        this.nextRequestId = dataManager.createReference(new ManagedSerializable<Integer>(1));
         this.callbacks = dataManager.createReference(new RemoteCallMap());
     }
 
-    private long getNextRequestId() {
-        final ManagedSerializable<Long> nextValue = nextRequestId.getForUpdate();
-        final long value = nextValue.get();
-        nextValue.set(value + 1L);
+    private int getNextRequestId() {
+        final ManagedSerializable<Integer> nextValue = nextRequestId.getForUpdate();
+        final int value = nextValue.get();
+        nextValue.set(value + 1);
         return value;
     }
 
     @Override
     public RemoteCall newRpcController() {
-        final long requestId = getNextRequestId();
+        final int requestId = getNextRequestId();
 
         final ManagedSerializable<RemoteCall> managedRemoteCall =
             new ManagedSerializable<RemoteCall>(new RemoteCallImpl(requestId));
@@ -47,7 +47,7 @@ public class ServerCallbackCache implements CallbackCache, Serializable {
     }
 
     @Override
-    public RemoteCall removeCallback(long requestId) {
+    public RemoteCall removeCallback(int requestId) {
         final ManagedSerializable<RemoteCall> managedCall = callbacks.getForUpdate().remove(requestId);
         AppContext.getDataManager().removeObject(managedCall);
         return managedCall.get();
